+++ tests/testthat/test-paths.R
@@ -198,22 +198,4 @@ test_that("zip file with non-ASCII characters", {
 
   unlink(zipfile)
 
-  # ----------------------------------------------------------------
-
-  p <- zippr::zip_process()$new(zipfile, c("dir1", "dir2"))
-  p$wait(5000)
-  p$kill()
-  expect_equal(p$get_exit_status(), 0L)
-
-  l <- zip_list(zipfile)
-  expect_equal(l$filename, c("dir1/", "dir1/file1", "dir2/", "dir2/file2"))
-
-  p2 <- zippr::unzip_process()$new(zipfile, "ex3")
-  p2$wait(5000)
-  p2$kill()
-  expect_equal(p2$get_exit_status(), 0L)
-  expect_equal(
-    sort(dir("ex3", recursive = TRUE)),
-    c("dir1/file1", "dir2/file2")
-  )
 })